Reintroduce component for sign-up.
